updated tests to check v3 intervention create api
