+++ src/etools/applications/partners/tests/test_api_interventions.py
@@ -33,7 +33,6 @@ from etools.applications.partners.tests.factories import (
     InterventionFactory,
     InterventionResultLinkFactory,
     PartnerFactory,
-    PartnerStaffFactory,
 )
 from etools.applications.partners.tests.test_utils import setup_intervention_test_data
 from etools.applications.reports.models import AppliedIndicator, ReportingRequirement
@@ -726,27 +725,6 @@ class TestInterventionsAPI(BaseTenantTestCase):
         )
         self.assertEqual(response.status_code, status.HTTP_200_OK)
         self.assertEqual(len(response.data), 1)
-
-    def test_add_intervention_by_partner_member(self):
-        partner_user = UserFactory(is_staff=False, groups__data=[])
-        PartnerStaffFactory(email=partner_user.email)
-        status_code, response = self.run_request_list_ep(user=partner_user, method='post', data={})
-        self.assertEqual(status_code, status.HTTP_403_FORBIDDEN, response)
-
-    def test_add_intervention_by_anonymous(self):
-        status_code, response = self.run_request_list_ep(user=AnonymousUser(), method='post', data={})
-        self.assertEqual(status_code, status.HTTP_403_FORBIDDEN, response)
-
-    def test_add_minimal_intervention(self):
-        status_code, response = self.run_request_list_ep(
-            user=self.partnership_manager_user, method='post',
-            data={
-                'document_type': Intervention.PD,
-                'title': 'My test intervention',
-                'agreement': self.agreement.id,
-            }
-        )
-        self.assertEqual(status_code, status.HTTP_201_CREATED, response)
 
 
 class TestAPIInterventionResultLinkListView(BaseTenantTestCase):
