+++ src/etools/applications/partners/tests/test_v3_interventions.py
@@ -1,5 +1,6 @@
 import datetime
 
+from django.contrib.auth.models import AnonymousUser
 from django.test import SimpleTestCase
 from django.urls import reverse
 
@@ -14,6 +15,7 @@ from etools.applications.partners.tests.factories import (
     InterventionFactory,
     InterventionResultLinkFactory,
     PartnerFactory,
+    PartnerStaffFactory,
 )
 from etools.applications.reports.models import ResultType
 from etools.applications.reports.tests.factories import (
@@ -95,6 +97,39 @@ class TestCreate(BaseInterventionTestCase):
         self.assertTrue(data.get("humanitarian_flag"))
         self.assertEqual(data.get("cfei_number"), "321")
         self.assertEqual(data.get("budget_owner"), self.user.pk)
+
+    def test_add_intervention_by_partner_member(self):
+        partner_user = UserFactory(is_staff=False, groups__data=[])
+        PartnerStaffFactory(email=partner_user.email)
+        response = self.forced_auth_req(
+            "post",
+            reverse('pmp_v3:intervention-list'),
+            user=partner_user,
+            data={}
+        )
+        self.assertEqual(response.status_code, status.HTTP_403_FORBIDDEN, response.data)
+
+    def test_add_intervention_by_anonymous(self):
+        response = self.forced_auth_req(
+            "post",
+            reverse('pmp_v3:intervention-list'),
+            user=AnonymousUser(),
+            data={}
+        )
+        self.assertEqual(response.status_code, status.HTTP_403_FORBIDDEN, response.data)
+
+    def test_add_minimal_intervention(self):
+        response = self.forced_auth_req(
+            "post",
+            reverse('pmp_v3:intervention-list'),
+            user=self.user,
+            data={
+                'document_type': Intervention.PD,
+                'title': 'My test intervention',
+                'agreement': self.agreement.pk,
+            }
+        )
+        self.assertEqual(response.status_code, status.HTTP_201_CREATED, response.data)
 
 
 class TestUpdate(BaseInterventionTestCase):
